Style: moved in_docker() to funcmodule
