+++ runner/funcmodule.py
@@ -20,7 +20,22 @@ text file.
 import click
 import pathlib
 import sys
+import os
 import yaml
+
+def in_docker() -> bool:
+    """Checks if code is currently running in a Docker container.
+    Checks if Docker is in control groups or if there is a `.dockerenv`
+    file at the filesystem's root directory.
+    Returns:
+        bool: Whether or not the code is running in a Docker container.
+    """
+    path = "/proc/self/cgroup"
+    return (
+        os.path.exists("/.dockerenv")
+        or os.path.isfile(path)
+        and any("docker" in line for line in open(path))
+    )
 
 def config_not_found_routine(config_file_path: pathlib.Path, DATA_DIR: pathlib.Path) -> None:
     """
